cond rendering for search on nav and modularized shipwreck array
